+++ reana-ui/src/pages/workflowDetails/WorkflowDetails.js
@@ -2,7 +2,7 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020, 2022 CERN.
+  Copyright (C) 2020, 2022, 2023 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
@@ -22,7 +22,11 @@ import {
   getConfig,
 } from "~/selectors";
 import BasePage from "../BasePage";
-import { Notification, WorkflowDeleteModal } from "~/components";
+import {
+  Notification,
+  WorkflowDeleteModal,
+  WorkflowStopModal,
+} from "~/components";
 import {
   WorkflowInfo,
   WorkflowLogs,
@@ -143,6 +147,7 @@ export default function WorkflowDetails() {
           defaultActiveIndex={defaultActiveIndex}
         />
         <WorkflowDeleteModal />
+        <WorkflowStopModal />
       </Container>
     </BasePage>
   );